requirements.txt
langchain==0.1.13
langchain_community==0.0.29
langchain_google_genai==1.0.1
protobuf==3.20.3
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.30.0
streamlit==1.32.2
